cleanup testing in lib
