+++ Cargo.toml
@@ -10,17 +10,17 @@ resolver = "2"
 [workspace.dependencies]
 alto-client = { version = "0.0.6", path = "client" }
 alto-types = { version = "0.0.6", path = "types" }
-commonware-codec = { version = "0.0.45" }
-commonware-consensus = { version = "0.0.45" }
-commonware-cryptography = { version = "0.0.45" }
-commonware-deployer = { version = "0.0.45" }
-commonware-macros = { version = "0.0.45" }
-commonware-p2p = { version = "0.0.45" }
-commonware-resolver = { version = "0.0.45" }
-commonware-runtime = { version = "0.0.45" }
-commonware-storage = { version = "0.0.45" }
-commonware-stream = { version = "0.0.45" }
-commonware-utils = { version = "0.0.45" }
+commonware-codec = { version = "0.0.45", path = "../monorepo/codec" }
+commonware-consensus = { version = "0.0.45", path = "../monorepo/consensus" }
+commonware-cryptography = { version = "0.0.45", path = "../monorepo/cryptography" }
+commonware-deployer = { version = "0.0.45", path = "../monorepo/deployer" }
+commonware-macros = { version = "0.0.45", path = "../monorepo/macros" }
+commonware-p2p = { version = "0.0.45", path = "../monorepo/p2p" }
+commonware-resolver = { version = "0.0.45", path = "../monorepo/resolver" }
+commonware-runtime = { version = "0.0.45", path = "../monorepo/runtime" }
+commonware-storage = { version = "0.0.45", path = "../monorepo/storage" }
+commonware-stream = { version = "0.0.45", path = "../monorepo/stream" }
+commonware-utils = { version = "0.0.45", path = "../monorepo/utils" }
 thiserror = "2.0.12"
 bytes = "1.7.1"
 rand = "0.8.5"
